+++ lib/feature/order/presentation/widgets/order_paid_tab.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:office_shopping_mall/feature/cart/data/order_list.dart';
-import 'package:office_shopping_mall/feature/cart/data/order_provider.dart';
+import 'package:office_shopping_mall/feature/order/data/order_list.dart';
+import 'package:office_shopping_mall/feature/order/data/order_provider.dart';
 
 class OrderPaidTab extends StatefulWidget{
   @override
