+++ lib/feature/order/presentation/widgets/order_shipping_tab.dart
@@ -1,16 +1,16 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:office_shopping_mall/feature/cart/data/order_list.dart';
-import 'package:office_shopping_mall/feature/cart/data/order_provider.dart';
+import 'package:office_shopping_mall/feature/order/data/order_list.dart';
+import 'package:office_shopping_mall/feature/order/data/order_provider.dart';
 
-class OrderDeliveredTab extends StatefulWidget{
+class OrderShippingTab extends StatefulWidget{
   @override
   State<StatefulWidget> createState() {
-    return OrderDeliveredTabState();
+    return OrderShippingTabState();
   }
 }
 
-class OrderDeliveredTabState extends State<OrderDeliveredTab> with SingleTickerProviderStateMixin{
+class OrderShippingTabState extends State<OrderShippingTab> with SingleTickerProviderStateMixin{
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -19,7 +19,7 @@ class OrderDeliveredTabState extends State<OrderDeliveredTab> with SingleTickerP
           orders: context
               .watch<OrderProvider>()
               .orders
-              .where((a) => a.orderIndex == 3)
+              .where((a) => a.orderIndex == 2)
               .toList(),
         ),
       ),
