+++ lib/main.dart
@@ -2,9 +2,9 @@ import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
 import 'package:office_shopping_mall/core/constants/app_constants.dart';
 import 'package:office_shopping_mall/core/utils/bottom_nav_observer.dart';
-import 'package:office_shopping_mall/feature/cart/data/order_provider.dart';
 import 'package:office_shopping_mall/core/providers/bottom_nav_provider.dart';
 import 'package:office_shopping_mall/feature/cart/presentation/cart_module.dart';
+import 'package:office_shopping_mall/feature/order/data/order_provider.dart';
 import 'package:office_shopping_mall/feature/product/product_module.dart';
 import 'package:office_shopping_mall/feature/search/search_module.dart';
 import 'package:office_shopping_mall/feature/setting/setting_module.dart';
